mythmediaserver: rename commandlineparser.* to mythmediaserver_…
